Download Saalfeld Lab registrations

Download Saalfeld Lab registrations

download_saalfeldlab_registrations(fileformat = c(".h5", ".nii"))

Arguments

fileformat
whether to download h5 (Saalfeld) or nii (ANTs) format registrations (defaults to h5).

Details

Registrations will be downloaded from https://www.janelia.org/open-science/jrc-2018-brain-templates. They will be downloaded to a folder defined by the package option nat.templatebrains.regdirs (see examples).

The h5 registration format typically provides a multi-resolution forward and inverse transformations in an efficiently compressed format. See https://github.com/saalfeldlab/template-building/wiki/Hdf5-Deformation-fields for further details.

The nii (NIfTI) files are the raw output format of the ANTs registration package. Their disadvantages are that they are 1) large files and the whole >500 Mb registration file must be loaded completely into memory before any transformations are possible, 2) that they depend on a more complex set of packages that can be hard to install, 3) is not simple to downsample to make smaller files. They do have the advantage of support for transforming image data as well as 3D points.

See also

Examples

regroot=getOption('nat.templatebrains.regdirs')
dir(regroot)

#>  [1] "ATAG_mirror.list"         "autoflip.sh"             
#>  [3] "Cell07_IS2.list"          "Dmel_DsecI.list"         
#>  [5] "Dmel_Dsim.list"           "Dmel_Dvir.list"          
#>  [7] "Dmel_Dyak.list"           "Dmel_IS2.list"           
#>  [9] "Dmel_JFRC2.list"          "doc"                     
#> [11] "doc"                      "DsecI_DsecF.list"        
#> [13] "DsecI_DsecM.list"         "DsecI_IS2.list"          
#> [15] "DsecI_JFRC2.list"         "Dsim_DsecI.list"         
#> [17] "FAFB14_FAFB14um.rds"      "FAFB14um_JRC2018F.rds"   
#> [19] "FCWB_IS2.list"            "FCWB_IS2.list"           
#> [21] "FCWB_JFRC2.list"          "FCWB_JFRC2.list"         
#> [23] "FCWB_JRC2018F.rds"        "FCWB_mirror.list"        
#> [25] "FCWB_mirror.list"         "IBNWB_IBN.list"          
#> [27] "IBNWB_IBN.list"           "IS2_Cell07.list"         
#> [29] "IS2_DsecI.list"           "IS2_mirror.list"         
#> [31] "IS2_mirror.list"          "IS2_T1.list"             
#> [33] "JFRC_mirror.list"         "JFRC2_DsecI.list"        
#> [35] "JFRC2_FCWB.list"          "JFRC2_IBNWB.list"        
#> [37] "JFRC2_IBNWB.list"         "JFRC2_imgflip.list"      
#> [39] "JFRC2_IS2.list"           "JFRC2_IS2.list"          
#> [41] "JFRC2_JFRC2013.list"      "JFRC2_mirror.list"       
#> [43] "JFRC2_mirror.list"        "JFRC2013_JFRC2013DS.list"
#> [45] "JFRC2013_JFRC2014.list"   "JFRC2013_JRC2018F.rds"   
#> [47] "JFRC2013_mirror.list"     "JFRC2013DS_JFRC2013.list"
#> [49] "JFRC2014_JFRC2014DS.list" "JFRC2014DS_JFRC2014.list"
#> [51] "JRC2018F_FAFB14um.rds"    "JRC2018F_FCWB.rds"       
#> [53] "JRC2018F_JFRC2013.rds"    "JRC2018F_JRCFIB2018F.rds"
#> [55] "JRCFIB2018F_JRC2018F.rds" "README.md"               
#> [57] "README.md"                "README.md"               
#> [59] "T1_FCWB.list"             "T1_IS2.list"             
#> [61] "T1_JFRC2.list"            "T1_mirror.list"          
#> [63] "T1_mirror.list"           "V2_mirror.list"          
#> [65] "VNCIS1_mirror.list"       "VNCIS1_V2.list"